% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahs_decision_tree.R
\name{ahs_decision_tree}
\alias{ahs_decision_tree}
\title{Average Household Size Decision Tree}
\usage{
ahs_decision_tree(
  external_ahs,
  trend_ahs,
  ahs_cap,
  projection_year,
  ldd_final_yr,
  gss_code_list
)
}
\arguments{
\item{external_ahs}{A dataframe of AHS values from the ONS or DCLG household model}

\item{trend_ahs}{A dataframe of AHS values from calculated from the
current projection year's trend population}

\item{ahs_cap}{A dataframe or NULL. If the cap has been set then a dataframe of
cap values. If the cap hasn't been set then NULL.}

\item{projection_year}{Numeric. The projection year}

\item{ldd_final_yr}{Numeric. The final year of data available in the LDD}

\item{gss_code_list}{Character. A vector of gss_codes which need to have an
AHS value returned}
}
\value{
A list where element 1 is the choice the tree made and element 2 is a
  dataframe of AHS values
}
\description{
A function to decide which of 3 AHS values to use: an external input value,
a capped value or a value consistent with trending the population forward
}
