% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_housingled_excels.R
\name{create_housingled_excels}
\alias{create_housingled_excels}
\title{Create the 2020-based housing-led projection excel output}
\usage{
create_housingled_excels(output_dir, wb_filename, projection_name)
}
\arguments{
\item{output_dir}{string. The directory from which to read the data and save the Excel file}

\item{wb_filename}{String. The name of the output Excel file. With or without '.xlsx'}

\item{projection_name}{String. The projection name.}
}
\description{
Process population, components and development data and place it into an
Excel template. London borough data.
}
