% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_led_core.R
\name{housing_led_core}
\alias{housing_led_core}
\title{Run a housing-led population model and output population, components of change and rates}
\usage{
housing_led_core(
  start_population,
  trend_projection,
  component_constraints,
  hma_constraint,
  communal_establishment_population,
  external_ahs,
  households_1,
  households_2,
  hma_list,
  projection_year,
  ahs_cap_year,
  ahs_cap,
  ahs_method,
  ldd_final_yr,
  constrain_projection,
  actual_births
)
}
\arguments{
\item{start_population}{A data frame. The population at the start of the time period}

\item{trend_projection}{A list. The output of the \code{trendmodel::trend_core}
function for the \code{projection_year}}

\item{component_constraints}{A dataframe. The component constraints by borough}

\item{communal_establishment_population}{A dataframe. The communal establishment
population for the \code{projection_year}}

\item{external_ahs}{A set of average household size rates by geographic
aggreagtion and year consistent with the \code{commual_establishment_population}}

\item{households_1}{A dataframe. For each geogrpahic aggregation a number of
total households to be used in calculating the current year trend ahs}

\item{households_2}{A dataframe. For each geogrpahic aggregation a number of
total households to be used in calculating the population by applying an AHS
rate}

\item{hma_list}{A named list. Containing gss_codes which, when aggregated,
form a housing market area (HMA). The name in the list is the name of the HMA.}

\item{projection_year}{Numeric. The year being projected}

\item{ahs_cap_year}{Numeric. The year in which to set the AHS cap}

\item{ahs_cap}{A dataframe or NULL. When no ahs cap has been set this
parameter is set to \code{NULL}. Once a set of ahs caps have been
this becomes the dataframe containing those caps}

\item{ahs_method}{A string or Numeric. If the ahs tree method is to be used,
set to \code{tree}, else the flaoting ahs method will be used. In this case
the parameter is a numeric between 0 and 1. It acts to determine the mix
between the external ahs and the trend ahs where 0 is the external and 1 is
the trend.}

\item{ldd_final_yr}{Numeric. The final year for which LDD development data 
is available.}

\item{constrain_projection}{Logical. If \code{TRUE} then the projection will
be constrained (at the borough level for components and at hma level for
population totals.) If \code{FALSE} the model is run unconstrained.}

\item{hma_constraint.}{A dataframe. The total population constrain by housing
market area (HMA)}
}
\value{
A list where each element is a data frame containing either projected population or
projected components of change.
}
\description{
Given a starting population and set of projected rates for fertility, mortality and migration
the function will produce a population for the next year using the GLA housing-led
method. The model outputs are the calculated populations and components of change.
}
