% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_borough_excels.R
\name{output_borough_excels}
\alias{output_borough_excels}
\title{Create the 2020-based housing-led projection excel output}
\usage{
output_borough_excels(output_dir, projection_name, file_suffix = ".xlsx")
}
\arguments{
\item{output_dir}{The directory from which to read the data and save the Excel file}

\item{projection_name}{The name of the output Excel file}

\item{file_suffix}{A string to append to the end of the file name}
}
\description{
Process population, components and development data and place it into an
Excel template. London borough data.
}
