% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_housing_led_excel_files.R
\name{output_housing_led_excel_file}
\alias{output_housing_led_excel_file}
\title{Create the 2019-based projection excel output}
\usage{
output_housing_led_excel_file(
  data,
  output_dir,
  projection_name,
  popn_adjustment_path,
  file_suffix = "_2019.xlsx"
)
}
\arguments{
\item{data}{The \code{[['csv']]} element from a ward projection}

\item{output_dir}{The directory in which to save the Excel file}

\item{projection_name}{The name of the output Excel file}

\item{popn_adjustment_path}{path to the population adjustment RDS file}

\item{file_suffix}{A string to append to the end of the file name}
}
\description{
Process population, components and development data and place it into an
Excel template.
}
