% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_housing_led_projection.R
\name{output_housing_led_projection}
\alias{output_housing_led_projection}
\title{Output the results of the housing-led model}
\usage{
output_housing_led_projection(
  projection,
  output_dir,
  external_trend_path,
  additional_dwellings,
  housing_stock,
  household_trajectory,
  first_proj_yr
)
}
\arguments{
\item{projection}{A list. The output from the run_housing_led_model function}

\item{output_dir}{A string. The directory in which to save the model output files}

\item{external_trend_path}{A string. The output path of the trend model used to
run the housing-led model. This is uses to import the backseries
components.}

\item{additional_dwellings}{A dataframe. The dwelling trajectory used in the
projection model.}

\item{housing_stock}{A dataframe. The total dwelling stock by year.}

\item{household_trajectory}{A dataframe. The household trajectory (as opposed
to the dwelling trajectory) used in the model.}

\item{first_proj_yr}{Numeric. The first year of the projection}
}
\description{
Save RDS and CSV files to a specified folder path. Outputs include population,
components of change (including backseries data) and the input dwelling and
household trajectories.
}
