% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_borough_and_ward_projection.R
\name{run_borough_and_ward_projection}
\alias{run_borough_and_ward_projection}
\title{Run the housing led and ward models with parametrised inputs}
\usage{
run_borough_and_ward_projection(
  projection_name,
  dev_trajectory_path,
  small_area_dev_trajectory_path,
  first_proj_yr,
  last_proj_yr,
  bpo = FALSE,
  housing_led_params = list(),
  small_area_params = list(),
  fertility_scenario,
  ldd_final_yr,
  constrain_projection
)
}
\arguments{
\item{projection_name}{String containing a name for the run. Used in the
output directory.}

\item{dev_trajectory_path}{String containing a path to the input development
trajectory.}

\item{small_area_dev_trajectory_path}{String containing a path the the input
development trajectory at small-area resolution.}

\item{housing_led_params, small_area_params}{Lists containing other key-value
pairs of configuration parameters for the two models. These will be written
over the default values, but won't supersede the above parameters. Note
that parameters constructed from other parameters, e.g.
external_ahs_trajectory_path which depends on external_trend_path, will
need to be specified as well.}

\item{first_proj_year}{Integer. First year of projection.}

\item{last_proj_year}{Integer. Last year of projection.}
}
\description{
Run the housing led and ward models with parametrised inputs
}
