% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpo_config_Jobscript.R
\name{run_bpo_config_Jobscript}
\alias{run_bpo_config_Jobscript}
\title{Run 3 concurrent BPO projections from one config file}
\usage{
run_bpo_config_Jobscript(config_name)
}
\arguments{
\item{config_name}{The name of the config file saved in the
\code{config_scripts/housing_led/bpo_conigs/} folder}
}
\description{
Use \code{rstudioapi::jobRunScript} to run 3 migration scenario projections
at the same time from a single BPO config script.
}
