% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_housing_led_model.R
\name{run_housing_led_model}
\alias{run_housing_led_model}
\title{Run a housing-led model config file to produce a population projection
Previously housing-led control}
\usage{
run_housing_led_model(config_list)
}
\arguments{
\item{config_list}{A List. A housing-led model configuration list.}
}
\description{
Read in, validate and manage model input data and then run the \code{trend_core}
and \code{housing_led_core} functions to produce a population projection as
specified in the config list.
}
