% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_bpo_to_datastore.R
\name{upload_bpo_to_datastore}
\alias{upload_bpo_to_datastore}
\title{Upload a BPO excel file to the datastore}
\usage{
upload_bpo_to_datastore(
  bpo_name,
  variants = c("high_migration", "medium_migration", "low_migration",
    "medium_migration_trend_fertility"),
  bpo_root = "outputs/housing_led/2018/bpo/"
)
}
\arguments{
\item{bpo_name}{String. The folder name within the \code{bpo_root} folder that
contains the file to be uploaded}

\item{variants}{Character. The projection variants. Defaults to \code{c("high_migration","medium_migration", "low_migration","medium_migration_trend_fertility")}}

\item{bpo_root}{String. The directory path to the bpo outputs folder.
Default \code{outputs/housing_led/2018/bpo/}}
}
\description{
Using the ldndatar package to leverage the datastore api
}
