% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_housing_led_core_outputs.R
\name{arrange_housing_led_core_outputs}
\alias{arrange_housing_led_core_outputs}
\title{Arrange outputs from the \code{housing_led_core} function into inputs for the
\code{output_housing_led_projections} function}
\usage{
arrange_housing_led_core_outputs(
  projection,
  trend_projection,
  first_proj_yr,
  last_proj_yr
)
}
\arguments{
\item{projection}{A list. The output list from the \code{housing_led_core} function}

\item{first_proj_yr}{Numeric. First projection year}

\item{last_proj_yr}{Numeric. Last projection year}

\item{trend_rpojection}{A list. The output flist rom the \code{trend_core} function}
}
\value{
A list where each element is a data frame containing data for each year
  of the projection and the backseries.
}
\description{
The \code{housing_led_core} outputs a list for each year of the projection.
Each list is a list of components. This function rearrages those outputs into
a list of dataframes of components.
}
