% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpo_projection.R
\name{run_bpo_projection}
\alias{run_bpo_projection}
\title{Run the borough and ward housing-led models using a BPO dwelling trajectory}
\usage{
run_bpo_projection(
  bpo_name,
  trajectory_range = 2020:2041,
  projection_range = 2020:2050,
  ldd_final_yr = 2019,
  bpo_dir = "Q:/Teams/D&PA/Demography/Projections/bpo_2019_based/",
  variant,
  csv_name = bpo_name,
  housing_led_params = list(),
  fertility_scenario = "trend"
)
}
\arguments{
\item{bpo_name}{String. The name of the dwelling trajectory csv saved in the
folder \code{bpo_dir} folder. The output projection will also have this name.}

\item{trajectory_range}{Numeric or NULL. The years covered by the input trajectory.
\code{Default 2020:2041}. If NULL the SHLAA trajectory is used.}

\item{projection_range}{Numeric. The years to project. \code{Default 2020:2050}.}

\item{ldd_final_yr}{Numeric. The final year of the LDD dwellings backseries.
\code{Default 2019}.}

\item{bpo_dir}{String. The folder containing the dwelling trajectory csv}

\item{variant}{String. The domestic migration scenario \code{high},
\code{high} or \code{low}.}

\item{csv_name}{String. The name of the dwelling trajectory csv saved in the
\code{bpo_dir} folder. With or without the "csv" suffix. Default is \code{bpo_name}}

\item{housing_led_params}{A list of parameters (a partial config list) to be
passed to the housing-led model. These will overwrite the default parameters
for the bpo projection.}

\item{fertility_scenario}{String. Should the model use fertility rates based
on an \code{average} or a \code{trend} of past data. Default is \code{average}}
}
\description{
A wrapper for run_borough_and_ward_projection. Takes as input a csv in the
of the GLA housing trajectory template for any borough and converts it into
an rds in tidy format. Run the borough housing-led model and the ward small
area model using a set of standard model parameters.
}
