% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpo_projection.R
\name{run_bpo_projection}
\alias{run_bpo_projection}
\title{Run the borough and ward housing-led models using a BPO dwelling trajectory}
\usage{
run_bpo_projection(
  bpo_name,
  shlaa_first_yr = 2042,
  first_proj_yr = 2019,
  final_proj_yr = 2050,
  dev_first_yr = 2019,
  ldd_final_yr = 2018,
  bpo_dir = "Q:/Teams/D&PA/Demography/Projections/bpo_2018_based/",
  migration_scenario
)
}
\arguments{
\item{bpo_name}{String. The name of the dwelling trajectory csv saved in the folder
\code{bpo_dir} folder. The output projection will also have this name.}

\item{shlaa_first_yr}{Numeric. The first year to use shlaa development data.
Effectively the final year of the supplied trajectory plus 1. \code{Default 2042}.}

\item{first_proj_yr}{Numeric. The first projection year \code{default 2019}}

\item{final_proj_yr}{Numeric. The final projection year \code{default 2050}}

\item{dev_first_yr}{Numeric. The first year for which development data is provided. \code{Default 2019}.}

\item{bpo_dir}{String. The folder containing the dwelling trajectory csv}

\item{ldd_last_yr}{Numeric. The final year of the LDD dwellings backsries. \code{Default 2018}.}

\item{migration_sceanrio}{String. The domestic migration scenario \code{high}, \code{medium} or \code{low}.}
}
\description{
A wrapper for create_bpo_trajectory and run_borough_and_ward
}
